backend/requirements.txt
asgiref==3.10.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
